simplification of model fetching
